+++ sqlite/table.go
@@ -38,6 +38,38 @@ func (t *sqliteConfig) ReadTableList() ([]string, error) {
 	return output, err
 }
 
+func (t *sqliteConfig) ReadCreateTableCmd(tname string) (string, error) {
+	var output string
+	cmd, err := readCreateTableCmd(tname)
+	if err != nil {
+		return output, err
+	}
+	rows, err := t.db.Query(cmd)
+	if err != nil {
+		return output, err
+	}
+	defer rows.Close()
+	for rows.Next() {
+		err = rows.Scan(&output)
+		if err != nil {
+			return "", err
+		}
+	}
+
+	return output, err
+
+}
+
+func (t *sqliteConfig) DropTable(tname string) error {
+	cmd, err := dropTableCmd(tname)
+	if err != nil {
+		return err
+	}
+	_, err = t.db.Exec(cmd)
+	return err
+
+}
+
 func createTableCmd(tname string, stu interface{}) (string, error) {
 	cmd := "CREATE TABLE IF NOT EXISTS" + " "
 	if tname == "" {
@@ -81,8 +113,19 @@ func createTableCmd(tname string, stu interface{}) (string, error) {
 
 }
 
-func readTableAllCmd() (string, error) {
-	cmd := "select name from sqlite_master where type='table'"
+func dropTableCmd(tname string) (string, error) {
+	cmd := "DROP TABLE IF EXISTS" + " '" + tname + "'"
 	return cmd, nil
 
+}
+
+func readTableAllCmd() (string, error) {
+	cmd := "SELECT name FROM sqlite_master WHERE type='table'"
+	return cmd, nil
+
+}
+
+func readCreateTableCmd(tname string) (string, error) {
+	cmd := "SELECT sql FROM sqlite_master WHERE type='table' AND name='" + tname + "'"
+	return cmd, nil
 }